+++ requirements.txt
@@ -2,3 +2,4 @@ requests
 beautifulsoup4
 lxml
 pytest
+pytz
